feat: Add new proto message 'ExtTrack' for track gate extension
Added a new proto message called 'ExtTrack' to the 'ext_track.proto' file in the 'junction/trackgate' package. The 'ExtTrack' message contains the fields 'trackName', 'trackId', 'trackKey', 'version', 'schema', and 'creators'.

This message is used to represent extended track information in the track gate module. It will be used to store and retrieve track information in the track gate module.

The 'ExtTrack' message has been added to the 'types' package in the 'junction/x/trackgate/types/message_track_creation.go.